refact : 로그아웃 성공 상태코드 수정 및 로직 수정
Redis 에 삭제 할 refreshToken 유무 확인 후 삭제 진행

20230912 이호찬

+++ user/src/main/java/com/jrjr/inbest/login/service/LoginServiceImpl.java
@@ -73,6 +73,8 @@ public class LoginServiceImpl implements LoginService {
 		}
 
 		// redis 에서 refreshToken 삭제
-		refreshTokenRepository.deleteById(inputLoginDto.getEmail());
+		if (refreshTokenRepository.existsById(inputLoginDto.getEmail())) {
+			refreshTokenRepository.deleteById(inputLoginDto.getEmail());
+		}
 	}
 }
